+++ cluster/cluster.go
@@ -6,13 +6,6 @@ import (
 	"google.golang.org/grpc"
 )
 
-type ClientConnInterface interface {
-	grpc.ClientConnInterface
-
-	Address() string
-	ID() uint32
-}
-
 type Cluster interface {
 	// ClientConnInterface interface allows Cluster use as grpc.ClientConnInterface
 	// with lazy getting raw grpc-connection in Invoke() or NewStream() stages.