+++ cluster/context.go
@@ -7,8 +7,7 @@ type (
 )
 
 type Endpoint interface {
-	Address() string
-	ID() uint32
+	NodeID() uint32
 }
 
 func WithEndpoint(ctx context.Context, endpoint Endpoint) context.Context {