+++ internal/table/session.go
@@ -3,6 +3,8 @@ package table
 import (
 	"context"
 	"io"
+	"net/url"
+	"strconv"
 	"sync"
 	"time"
 
@@ -14,7 +16,6 @@ import (
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Table"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/cluster"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/driver"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/feature"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/operation"
@@ -48,13 +49,24 @@ const (
 // Close() call.
 type session struct {
 	id           string
-	endpoint     cluster.Endpoint
 	tableService Ydb_Table_V1.TableServiceClient
 	trace        trace.Table
 	mtx          sync.Mutex
 	flags        sessionFlags
 	status       options.SessionStatus
 	onClose      []func(ctx context.Context)
+}
+
+func (s *session) NodeID() uint32 {
+	u, err := url.Parse(s.id)
+	if err != nil {
+		panic(err)
+	}
+	nodeID, err := strconv.ParseUint(u.Query().Get("node_id"), 10, 32)
+	if err != nil {
+		return 0
+	}
+	return uint32(nodeID)
 }
 
 func (s *session) Status() string {
@@ -103,16 +115,12 @@ func newSession(ctx context.Context, cc grpc.ClientConnInterface, t trace.Table)
 	var (
 		response *Ydb_Table.CreateSessionResponse
 		result   Ydb_Table.CreateSessionResult
-		endpoint cluster.Endpoint
 	)
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
 	c := Ydb_Table_V1.NewTableServiceClient(cc)
-	response, err = c.CreateSession(
-		driver.WithCallInfo(ctx, func(e cluster.Endpoint) { endpoint = e }),
-		&Ydb_Table.CreateSessionRequest{},
-	)
+	response, err = c.CreateSession(ctx, &Ydb_Table.CreateSessionRequest{})
 	if err != nil {
 		return nil, err
 	}
@@ -122,7 +130,6 @@ func newSession(ctx context.Context, cc grpc.ClientConnInterface, t trace.Table)
 	}
 	s = &session{
 		id:           result.GetSessionId(),
-		endpoint:     endpoint,
 		tableService: c,
 		trace:        t,
 	}
@@ -167,7 +174,7 @@ func (s *session) Close(ctx context.Context) (err error) {
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	_, err = s.tableService.DeleteSession(cluster.WithEndpoint(ctx, s.endpoint), &Ydb_Table.DeleteSessionRequest{
+	_, err = s.tableService.DeleteSession(cluster.WithEndpoint(ctx, s), &Ydb_Table.DeleteSessionRequest{
 		SessionId: s.id,
 	})
 	return err
@@ -183,7 +190,7 @@ func (s *session) KeepAlive(ctx context.Context) (err error) {
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	resp, err := s.tableService.KeepAlive(cluster.WithEndpoint(ctx, s.endpoint), &Ydb_Table.KeepAliveRequest{
+	resp, err := s.tableService.KeepAlive(cluster.WithEndpoint(ctx, s), &Ydb_Table.KeepAliveRequest{
 		SessionId: s.id,
 	})
 	if err != nil {
@@ -211,7 +218,7 @@ func (s *session) CreateTable(ctx context.Context, path string, opts ...options.
 	for _, opt := range opts {
 		opt((*options.CreateTableDesc)(&request))
 	}
-	_, err = s.tableService.CreateTable(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	_, err = s.tableService.CreateTable(cluster.WithEndpoint(ctx, s), &request)
 	return err
 }
 
@@ -228,7 +235,7 @@ func (s *session) DescribeTable(ctx context.Context, path string, opts ...option
 	for _, opt := range opts {
 		opt((*options.DescribeTableDesc)(&request))
 	}
-	response, err = s.tableService.DescribeTable(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	response, err = s.tableService.DescribeTable(cluster.WithEndpoint(ctx, s), &request)
 	if err != nil {
 		return desc, err
 	}
@@ -334,7 +341,7 @@ func (s *session) DropTable(ctx context.Context, path string, opts ...options.Dr
 	for _, opt := range opts {
 		opt((*options.DropTableDesc)(&request))
 	}
-	_, err = s.tableService.DropTable(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	_, err = s.tableService.DropTable(cluster.WithEndpoint(ctx, s), &request)
 	return err
 }
 
@@ -347,7 +354,7 @@ func (s *session) AlterTable(ctx context.Context, path string, opts ...options.A
 	for _, opt := range opts {
 		opt((*options.AlterTableDesc)(&request))
 	}
-	_, err = s.tableService.AlterTable(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	_, err = s.tableService.AlterTable(cluster.WithEndpoint(ctx, s), &request)
 	return err
 }
 
@@ -361,7 +368,7 @@ func (s *session) CopyTable(ctx context.Context, dst, src string, opts ...option
 	for _, opt := range opts {
 		opt((*options.CopyTableDesc)(&request))
 	}
-	_, err = s.tableService.CopyTable(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	_, err = s.tableService.CopyTable(cluster.WithEndpoint(ctx, s), &request)
 	return err
 }
 
@@ -374,7 +381,7 @@ func (s *session) Explain(ctx context.Context, query string) (exp table.DataQuer
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	response, err = s.tableService.ExplainDataQuery(cluster.WithEndpoint(ctx, s.endpoint), &Ydb_Table.ExplainDataQueryRequest{
+	response, err = s.tableService.ExplainDataQuery(cluster.WithEndpoint(ctx, s), &Ydb_Table.ExplainDataQueryRequest{
 		SessionId: s.id,
 		YqlText:   query,
 	})
@@ -452,7 +459,7 @@ func (s *session) Prepare(ctx context.Context, query string) (stmt table.Stateme
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	response, err = s.tableService.PrepareDataQuery(cluster.WithEndpoint(ctx, s.endpoint), &Ydb_Table.PrepareDataQueryRequest{
+	response, err = s.tableService.PrepareDataQuery(cluster.WithEndpoint(ctx, s), &Ydb_Table.PrepareDataQueryRequest{
 		SessionId: s.id,
 		YqlText:   query,
 	})
@@ -555,7 +562,7 @@ func (s *session) executeDataQuery(
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	response, err = s.tableService.ExecuteDataQuery(cluster.WithEndpoint(ctx, s.endpoint), request)
+	response, err = s.tableService.ExecuteDataQuery(cluster.WithEndpoint(ctx, s), request)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -572,7 +579,7 @@ func (s *session) ExecuteSchemeQuery(ctx context.Context, query string, opts ...
 	for _, opt := range opts {
 		opt((*options.ExecuteSchemeQueryDesc)(&request))
 	}
-	_, err = s.tableService.ExecuteSchemeQuery(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	_, err = s.tableService.ExecuteSchemeQuery(cluster.WithEndpoint(ctx, s), &request)
 	return err
 }
 
@@ -583,7 +590,7 @@ func (s *session) DescribeTableOptions(ctx context.Context) (desc options.TableO
 		result   Ydb_Table.DescribeTableOptionsResult
 	)
 	request := Ydb_Table.DescribeTableOptionsRequest{}
-	response, err = s.tableService.DescribeTableOptions(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	response, err = s.tableService.DescribeTableOptions(cluster.WithEndpoint(ctx, s), &request)
 	if err != nil {
 		return
 	}
@@ -697,7 +704,7 @@ func (s *session) StreamReadTable(ctx context.Context, path string, opts ...opti
 
 	ctx, cancel := context.WithCancel(ctx)
 
-	c, err = s.tableService.StreamReadTable(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	c, err = s.tableService.StreamReadTable(cluster.WithEndpoint(ctx, s), &request)
 
 	onDone := trace.TableOnSessionQueryStreamRead(s.trace, ctx, s)
 	if err != nil {
@@ -767,7 +774,7 @@ func (s *session) StreamExecuteScanQuery(ctx context.Context, query string, para
 
 	ctx, cancel := context.WithCancel(ctx)
 
-	c, err = s.tableService.StreamExecuteScanQuery(cluster.WithEndpoint(ctx, s.endpoint), &request)
+	c, err = s.tableService.StreamExecuteScanQuery(cluster.WithEndpoint(ctx, s), &request)
 
 	onDone := trace.TableOnSessionQueryStreamExecute(s.trace, ctx, s, q, params)
 	if err != nil {
@@ -820,7 +827,7 @@ func (s *session) StreamExecuteScanQuery(ctx context.Context, query string, para
 
 // BulkUpsert uploads given list of ydb struct values to the table.
 func (s *session) BulkUpsert(ctx context.Context, table string, rows types.Value) (err error) {
-	_, err = s.tableService.BulkUpsert(cluster.WithEndpoint(ctx, s.endpoint), &Ydb_Table.BulkUpsertRequest{
+	_, err = s.tableService.BulkUpsert(cluster.WithEndpoint(ctx, s), &Ydb_Table.BulkUpsertRequest{
 		Table: table,
 		Rows:  value.ToYDB(rows),
 	})
@@ -841,7 +848,7 @@ func (s *session) BeginTransaction(ctx context.Context, tx *table.TransactionSet
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	response, err = s.tableService.BeginTransaction(cluster.WithEndpoint(ctx, s.endpoint), &Ydb_Table.BeginTransactionRequest{
+	response, err = s.tableService.BeginTransaction(cluster.WithEndpoint(ctx, s), &Ydb_Table.BeginTransactionRequest{
 		SessionId:  s.id,
 		TxSettings: tx.Settings(),
 	})
@@ -915,7 +922,7 @@ func (tx *Transaction) CommitTx(ctx context.Context, opts ...options.CommitTrans
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	response, err = tx.s.tableService.CommitTransaction(cluster.WithEndpoint(ctx, tx.s.endpoint), request)
+	response, err = tx.s.tableService.CommitTransaction(cluster.WithEndpoint(ctx, tx.s), request)
 	if err != nil {
 		return nil, err
 	}
@@ -935,7 +942,7 @@ func (tx *Transaction) Rollback(ctx context.Context) (err error) {
 	if m, _ := operation.ContextMode(ctx); m == operation.ModeUnknown {
 		ctx = operation.WithMode(ctx, operation.ModeSync)
 	}
-	_, err = tx.s.tableService.RollbackTransaction(cluster.WithEndpoint(ctx, tx.s.endpoint), &Ydb_Table.RollbackTransactionRequest{
+	_, err = tx.s.tableService.RollbackTransaction(cluster.WithEndpoint(ctx, tx.s), &Ydb_Table.RollbackTransactionRequest{
 		SessionId: tx.s.id,
 		TxId:      tx.id,
 	})