+++ trace/driver.go
@@ -77,7 +77,7 @@ type ConnState interface {
 }
 
 type endpointInfo interface {
-	ID() uint32
+	NodeID() uint32
 	Address() string
 	LocalDC() bool
 }